requirements.txt
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
binance-futures-connector==3.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pyOpenSSL==23.1.0
requests==2.28.2
service-identity==21.1.0
six==1.16.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.15
zope.interface==6.0
